+++ app/src/main/java/com/orangeelephant/sobriety/storage/models/Reason.kt
@@ -2,6 +2,6 @@ package com.orangeelephant.sobriety.storage.models
 
 data class Reason(
     val id: Int,
-    val counterId: Int,
+    val counterId: Long,
     val reason: String
 )
