+++ app/src/main/java/com/orangeelephant/sobriety/storage/repositories/CounterRepository.kt
@@ -1,14 +1,13 @@
 package com.orangeelephant.sobriety.storage.repositories
 
 import com.orangeelephant.sobriety.storage.models.Counter
-import com.orangeelephant.sobriety.storage.models.Reason
 import com.orangeelephant.sobriety.storage.models.Relapse
 
 interface CounterRepository {
     fun getAllCounters(): List<Counter>
-    fun getCounter(id: Int): Counter
-    fun getRelapsesForCounter(counterId: Int): List<Relapse>
+    fun getCounter(id: Long): Counter
+    fun getRelapsesForCounter(counterId: Long): List<Relapse>
     fun addCounter(counter: Counter, list: List<String>): Long
-    fun resetCounter(id: Int, comment: String?): Long
-    fun deleteCounter(id: Int)
+    fun resetCounter(id: Long, comment: String?): Long
+    fun deleteCounter(id: Long)
 }
